refactor(component) : refactor SearchBar
- move SearchBar into common component directory
- rename SearchPainting component to distinguish SearchPaintingBar and SearchBar
